Move raster opacity to seperate example
