+++ examples/rasteropacity/index.html
@@ -1,0 +1,10 @@
+---
+layout: example.html
+title: Raster Opacity Example
+shortdesc: Demonstrates how to change a raster's opacity.
+docs: >
+  <li>Using raster opacity</li>
+
+---
+<div id="features"></div>
+<div id="controls"></div>
